third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGClipPathElement-svgdom-clipPathUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'clipPathUnits' property of the SVGClipPathElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS clipPathElement.clipPathUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS clipPathElement.clipPathUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS successfullyParsed is true

TEST COMPLETE
